+++ NEWS.md
@@ -1,6 +1,7 @@
 # map (development version)
 
 * drop support for R 4.0 due to changes in dependency packages
+* [WIP] remove retiring dependency packages `sp`, `rgeos` in favour of `terra` + `sf`
 
 # map v0.0.3