+++ NEWS.md
@@ -1,6 +1,8 @@
 # map (development version)
 
-* drop support for R 4.0 due to changes in dependency packages
+* drop support for R 4.0 due to changes in dependency packages;
+* use `digest::digest(x, algo = "spooky")` instead of `fastdigest` (which was archived on CRAN);
+* remove unused dependencies `knitr` and `rmarkdown`;
 
 # map v0.0.3